requirements.txt
flask
text2emotion
requests
python-dotenv
huggingface_hub